# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://localhost:8080/home
  description: Go Gin 框架示例项目 API 文档，包含用户认证、文章管理、标签管理等接口
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin Demo API
  version: "1.0"
paths:
...
  /user/check-email:
    get:
      description: 检查指定邮箱是否已被其他用户注册，用于注册时实时校验
      parameters:
      - description: 邮箱地址
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 检查结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 检查邮箱是否存在
      tags:
      - 用户认证
  /user/check-username:
    get:
      description: 检查指定用户名是否已被其他用户使用，用于注册时实时校验
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 检查结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 检查用户名是否存在
      tags:
      - 用户认证
  /user/login:
    get:
      description: 返回用户登录表单页面，支持用户名和邮箱两种登录方式
      responses:
        "200":
          description: 登录页面HTML
          schema:
            type: string
      summary: 登录页面
      tags:
      - 用户认证
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 通过用户名或邮箱进行登录，支持切换登录方式。登录成功后返回Token并设置Cookie。
      parameters:
      - description: 账号（用户名或邮箱）
        in: formData
        name: account
        required: true
        type: string
      - description: 登录类型：email 或 username（默认自动判断）
        in: formData
        name: login_type
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: 重定向到首页
          schema:
            type: string
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 账号或密码错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 用户登录
      tags:
      - 用户认证
  /user/logout:
    get:
      description: 清除当前用户的登录Token并重定向到登录页
      responses:
        "302":
          description: 重定向到登录页
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - 用户认证
  /user/register:
    get:
      description: 返回用户注册表单页面
      responses:
        "200":
          description: 注册页面HTML
          schema:
            type: string
      summary: 注册页面
      tags:
      - 用户认证
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 创建新用户账号，需要提供用户名、邮箱、密码（含确认密码）和邮箱验证码
      parameters:
      - description: 用户名（3-20位字母数字下划线）
        in: formData
        name: username
        required: true
        type: string
      - description: 邮箱地址
        in: formData
        name: email
        required: true
        type: string
      - description: 密码（至少8位，包含2种以上字符类型）
        in: formData
        name: password
        required: true
        type: string
      - description: 确认密码
        in: formData
        name: confirm_password
        required: true
        type: string
      - description: 邮箱验证码（6位数字）
        in: formData
        name: verify_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: 重定向到登录页
          schema:
            type: string
        "400":
          description: 参数错误或格式不合法
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 用户名或邮箱已存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户注册
      tags:
      - 用户认证
  /user/send-code:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 向指定邮箱发送6位数字验证码，用于注册验证。验证码5分钟内有效。
      parameters:
      - description: 邮箱地址
        in: formData
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 发送成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 发送失败
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 发送邮箱验证码
      tags:
      - 用户认证
securityDefinitions:
  BearerAuth:
    description: '请在请求头中携带 Token: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"